fix: parse nuget version
