+++ creation/constants.go
@@ -5,8 +5,6 @@
 package creation
 
 import (
-	"github.com/Masterminds/semver/v3"
-
 	"github.com/huhouhua/go-nuget"
 )
 
@@ -18,64 +16,64 @@ const (
 
 var (
 	// Net2 is net20 (.NETFramework,Version=v2.0)
-	Net2 = NewFrameworkWithVersion(nuget.Net, semver.New(2, 0, 0, "", ""))
+	Net2 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(2, 0, 0, "", ""))
 	// Net35 is net35 (.NETFramework,Version=v3.5)
-	Net35 = NewFrameworkWithVersion(nuget.Net, semver.New(3, 5, 0, "", ""))
+	Net35 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(3, 5, 0, "", ""))
 	// Net4 is net40 (.NETFramework,Version=v4.0)
-	Net4 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 0, 0, "", ""))
+	Net4 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 0, 0, "", ""))
 	// Net403 is net403 (.NETFramework,Version=v4.0.3)
-	Net403 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 0, 3, "", ""))
+	Net403 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 0, 3, "", ""))
 	// Net45 is net45 (.NETFramework,Version=v4.5)
-	Net45 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 5, 0, "", ""))
+	Net45 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 5, 0, "", ""))
 	// Net451 is net451 (.NETFramework,Version=v4.5.1)
-	Net451 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 5, 1, "", ""))
+	Net451 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 5, 1, "", ""))
 	// Net452 is net452 (.NETFramework,Version=v4.5.2)
-	Net452 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 5, 2, "", ""))
+	Net452 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 5, 2, "", ""))
 	// Net46 is net46 (.NETFramework,Version=v4.6)
-	Net46 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 6, 0, "", ""))
+	Net46 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 6, 0, "", ""))
 	// Net461 is net461 (.NETFramework,Version=v4.6.1)
-	Net461 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 6, 1, "", ""))
+	Net461 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 6, 1, "", ""))
 	// Net462 is net462 (.NETFramework,Version=v4.6.2)
-	Net462 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 6, 2, "", ""))
+	Net462 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 6, 2, "", ""))
 	// Net463 is net463 (.NETFramework,Version=v4.6.3)
-	Net463 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 6, 3, "", ""))
+	Net463 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 6, 3, "", ""))
 	// Net47 is net47 (.NETFramework,Version=v4.7)
-	Net47 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 7, 0, "", ""))
+	Net47 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 7, 0, "", ""))
 	// Net471 is net471 (.NETFramework,Version=v4.7.1)
-	Net471 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 7, 1, "", ""))
+	Net471 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 7, 1, "", ""))
 	// Net472 is net472 (.NETFramework,Version=v4.7.2)
-	Net472 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 7, 2, "", ""))
+	Net472 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 7, 2, "", ""))
 	// Net48 is net48 (.NETFramework,Version=v4.8)
-	Net48 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 8, 0, "", ""))
+	Net48 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 8, 0, "", ""))
 	// Net481 is net481 (.NETFramework,Version=v4.8.1)
-	Net481 = NewFrameworkWithVersion(nuget.Net, semver.New(4, 8, 1, "", ""))
+	Net481 = NewFrameworkWithVersion(nuget.Net, nuget.NewVersionFrom(4, 8, 1, "", ""))
 
 	// NetCore45 is netcore45 (.NETCore,Version=v4.5)
 	// This is not .NET Core. You are probably looking for netcoreapp.
-	NetCore45 = NewFrameworkWithVersion(nuget.NetCore, semver.New(4, 5, 0, "", ""))
+	NetCore45 = NewFrameworkWithVersion(nuget.NetCore, nuget.NewVersionFrom(4, 5, 0, "", ""))
 	// NetCore451 is netcore451 (.NETCore,Version=v4.5.1)
 	// This is not .NET Core. You are probably looking for netcoreapp.
-	NetCore451 = NewFrameworkWithVersion(nuget.NetCore, semver.New(4, 5, 1, "", ""))
+	NetCore451 = NewFrameworkWithVersion(nuget.NetCore, nuget.NewVersionFrom(4, 5, 1, "", ""))
 	// NetCore50 is netcore50 (.NETCore,Version=v5.0)
 	// This is not .NET 5. You are probably looking for net50 (.NETCoreApp,Version=v5.0)
-	NetCore50 = NewFrameworkWithVersion(nuget.NetCore, semver.New(5, 0, 0, "", ""))
+	NetCore50 = NewFrameworkWithVersion(nuget.NetCore, nuget.NewVersionFrom(5, 0, 0, "", ""))
 
-	Win8  = NewFrameworkWithVersion(nuget.Windows, semver.New(8, 0, 0, "", ""))
-	Win81 = NewFrameworkWithVersion(nuget.Windows, semver.New(8, 1, 0, "", ""))
-	Win10 = NewFrameworkWithVersion(nuget.Windows, semver.New(10, 0, 0, "", ""))
+	Win8  = NewFrameworkWithVersion(nuget.Windows, nuget.NewVersionFrom(8, 0, 0, "", ""))
+	Win81 = NewFrameworkWithVersion(nuget.Windows, nuget.NewVersionFrom(8, 1, 0, "", ""))
+	Win10 = NewFrameworkWithVersion(nuget.Windows, nuget.NewVersionFrom(10, 0, 0, "", ""))
 
-	SL4 = NewFrameworkWithVersion(nuget.Silverlight, semver.New(4, 0, 0, "", ""))
-	SL5 = NewFrameworkWithVersion(nuget.Silverlight, semver.New(5, 0, 0, "", ""))
+	SL4 = NewFrameworkWithVersion(nuget.Silverlight, nuget.NewVersionFrom(4, 0, 0, "", ""))
+	SL5 = NewFrameworkWithVersion(nuget.Silverlight, nuget.NewVersionFrom(5, 0, 0, "", ""))
 
-	WP7   = NewFrameworkWithVersion(nuget.WindowsPhone, semver.New(7, 0, 0, "", ""))
-	WP75  = NewFrameworkWithVersion(nuget.WindowsPhone, semver.New(7, 5, 0, "", ""))
-	WP8   = NewFrameworkWithVersion(nuget.WindowsPhone, semver.New(8, 0, 0, "", ""))
-	WP81  = NewFrameworkWithVersion(nuget.WindowsPhone, semver.New(8, 1, 0, "", ""))
-	WPA81 = NewFrameworkWithVersion(nuget.WindowsPhoneApp, semver.New(8, 1, 0, "", ""))
+	WP7   = NewFrameworkWithVersion(nuget.WindowsPhone, nuget.NewVersionFrom(7, 0, 0, "", ""))
+	WP75  = NewFrameworkWithVersion(nuget.WindowsPhone, nuget.NewVersionFrom(7, 5, 0, "", ""))
+	WP8   = NewFrameworkWithVersion(nuget.WindowsPhone, nuget.NewVersionFrom(8, 0, 0, "", ""))
+	WP81  = NewFrameworkWithVersion(nuget.WindowsPhone, nuget.NewVersionFrom(8, 1, 0, "", ""))
+	WPA81 = NewFrameworkWithVersion(nuget.WindowsPhoneApp, nuget.NewVersionFrom(8, 1, 0, "", ""))
 
-	Tizen3 = NewFrameworkWithVersion(nuget.Tizen, semver.New(3, 0, 0, "", ""))
-	Tizen4 = NewFrameworkWithVersion(nuget.Tizen, semver.New(4, 0, 0, "", ""))
-	Tizen6 = NewFrameworkWithVersion(nuget.Tizen, semver.New(6, 0, 0, "", ""))
+	Tizen3 = NewFrameworkWithVersion(nuget.Tizen, nuget.NewVersionFrom(3, 0, 0, "", ""))
+	Tizen4 = NewFrameworkWithVersion(nuget.Tizen, nuget.NewVersionFrom(4, 0, 0, "", ""))
+	Tizen6 = NewFrameworkWithVersion(nuget.Tizen, nuget.NewVersionFrom(6, 0, 0, "", ""))
 
 	AspNet       = NewFrameworkWithVersion(nuget.AspNet, nuget.EmptyVersion)
 	AspNetCore   = NewFrameworkWithVersion(nuget.AspNetCore, nuget.EmptyVersion)
@@ -83,61 +81,61 @@ var (
 	AspNetCore50 = NewFrameworkWithVersion(nuget.AspNetCore, nuget.Version5)
 
 	Dnx       = NewFrameworkWithVersion(nuget.Dnx, nuget.EmptyVersion)
-	Dnx45     = NewFrameworkWithVersion(nuget.Dnx, semver.New(4, 5, 0, "", ""))
-	Dnx452    = NewFrameworkWithVersion(nuget.Dnx, semver.New(4, 5, 2, "", ""))
+	Dnx45     = NewFrameworkWithVersion(nuget.Dnx, nuget.NewVersionFrom(4, 5, 0, "", ""))
+	Dnx452    = NewFrameworkWithVersion(nuget.Dnx, nuget.NewVersionFrom(4, 5, 2, "", ""))
 	DnxCore   = NewFrameworkWithVersion(nuget.DnxCore, nuget.EmptyVersion)
 	DnxCore50 = NewFrameworkWithVersion(nuget.DnxCore, nuget.Version5)
 
 	DotNet   = NewFrameworkWithVersion(nuget.NetPlatform, nuget.EmptyVersion)
 	DotNet50 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.Version5)
 
-	DotNet51 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 1, 0, "", ""))
-	DotNet52 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 2, 0, "", ""))
-	DotNet53 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 3, 0, "", ""))
-	DotNet54 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 4, 0, "", ""))
-	DotNet55 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 5, 0, "", ""))
-	DotNet56 = NewFrameworkWithVersion(nuget.NetPlatform, semver.New(5, 6, 0, "", ""))
+	DotNet51 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 1, 0, "", ""))
+	DotNet52 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 2, 0, "", ""))
+	DotNet53 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 3, 0, "", ""))
+	DotNet54 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 4, 0, "", ""))
+	DotNet55 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 5, 0, "", ""))
+	DotNet56 = NewFrameworkWithVersion(nuget.NetPlatform, nuget.NewVersionFrom(5, 6, 0, "", ""))
 
 	NetStandard = NewFrameworkWithVersion(nuget.NetStandard, nuget.EmptyVersion)
 	// NetStandard10 is netstandard1.0 (.NETStandard,Version=v1.0)
-	NetStandard10 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 0, 0, "", ""))
+	NetStandard10 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 0, 0, "", ""))
 	// NetStandard11 is netstandard1.1 (.NETStandard,Version=v1.1)
-	NetStandard11 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 1, 0, "", ""))
+	NetStandard11 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 1, 0, "", ""))
 	// NetStandard12 is netstandard1.2 (.NETStandard,Version=v1.2)
-	NetStandard12 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 2, 0, "", ""))
+	NetStandard12 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 2, 0, "", ""))
 	// NetStandard13 is netstandard1.3 (.NETStandard,Version=v1.3)
-	NetStandard13 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 3, 0, "", ""))
+	NetStandard13 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 3, 0, "", ""))
 	// NetStandard14 is netstandard1.4 (.NETStandard,Version=v1.4)
-	NetStandard14 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 4, 0, "", ""))
+	NetStandard14 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 4, 0, "", ""))
 	// NetStandard15 is netstandard1.5 (.NETStandard,Version=v1.5)
-	NetStandard15 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 5, 0, "", ""))
+	NetStandard15 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 5, 0, "", ""))
 	// NetStandard16 is netstandard1.6 (.NETStandard,Version=v1.6)
-	NetStandard16 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 6, 0, "", ""))
+	NetStandard16 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 6, 0, "", ""))
 	// NetStandard17 is netstandard1.7 (.NETStandard,Version=v1.7
-	NetStandard17 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(1, 7, 0, "", ""))
+	NetStandard17 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(1, 7, 0, "", ""))
 	// NetStandard20 is netstandard2.0 (.NETStandard,Version=v2.0)
-	NetStandard20 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(2, 0, 0, "", ""))
+	NetStandard20 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(2, 0, 0, "", ""))
 	// NetStandard21 is netstandard2.1 (.NETStandard,Version=v2.1)
-	NetStandard21 = NewFrameworkWithVersion(nuget.NetStandard, semver.New(2, 1, 0, "", ""))
+	NetStandard21 = NewFrameworkWithVersion(nuget.NetStandard, nuget.NewVersionFrom(2, 1, 0, "", ""))
 
-	NetStandardApp15 = NewFrameworkWithVersion(nuget.NetStandardApp, semver.New(1, 5, 0, "", ""))
+	NetStandardApp15 = NewFrameworkWithVersion(nuget.NetStandardApp, nuget.NewVersionFrom(1, 5, 0, "", ""))
 
 	UAP10 = NewFrameworkWithVersion(nuget.UAP, nuget.Version10)
 
 	// NetCoreApp10 is netcoreapp1.0 (.NETCoreApp,Version=v1.0)
-	NetCoreApp10 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(1, 0, 0, "", ""))
+	NetCoreApp10 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(1, 0, 0, "", ""))
 	// NetCoreApp11 is netcoreapp1.1 (.NETCoreApp,Version=v1.1)
-	NetCoreApp11 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(1, 1, 0, "", ""))
+	NetCoreApp11 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(1, 1, 0, "", ""))
 	// NetCoreApp20 is netcoreapp2.0 (.NETCoreApp,Version=v2.0)
-	NetCoreApp20 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(2, 0, 0, "", ""))
+	NetCoreApp20 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(2, 0, 0, "", ""))
 	// NetCoreApp21 is netcoreapp2.1 (.NETCoreApp,Version=v2.1)
-	NetCoreApp21 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(2, 1, 0, "", ""))
+	NetCoreApp21 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(2, 1, 0, "", ""))
 	// NetCoreApp22 is netcoreapp2.2 (.NETCoreApp,Version=v2.2)
-	NetCoreApp22 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(2, 2, 0, "", ""))
+	NetCoreApp22 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(2, 2, 0, "", ""))
 	// NetCoreApp30 is netcoreapp3.0 (.NETCoreApp,Version=v3.0)
-	NetCoreApp30 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(3, 0, 0, "", ""))
+	NetCoreApp30 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(3, 0, 0, "", ""))
 	// NetCoreApp31 is netcoreapp3.1 (.NETCoreApp,Version=v3.1)
-	NetCoreApp31 = NewFrameworkWithVersion(nuget.NetCoreApp, semver.New(3, 1, 0, "", ""))
+	NetCoreApp31 = NewFrameworkWithVersion(nuget.NetCoreApp, nuget.NewVersionFrom(3, 1, 0, "", ""))
 
 	// Net50 .NET 5.0 and later has NetCoreApp identifier
 	// Net50 is net5.0 (.NETCoreApp,Version=v5.0)
