+++ search.go
@@ -7,8 +7,6 @@ package nuget
 import (
 	"net/http"
 	"net/url"
-
-	"github.com/Masterminds/semver/v3"
 )
 
 type PackageSearchResource struct {
@@ -57,8 +55,8 @@ type VersionInfo struct {
 	DownloadCount uint64 `json:"downloads"`
 }
 
-func (v *VersionInfo) ParseVersion() (*semver.Version, error) {
-	ver, err := semver.NewVersion(v.Version)
+func (v *VersionInfo) ParseVersion() (*Version, error) {
+	ver, err := ParseVersion(v.Version)
 	if err != nil {
 		return nil, err
 	}
